requirements.txt
aiogram==2.25.1
ffmpeg==1.4
numpy==1.24.0
opencv-python==4.6.0.66
Pillow==9.3.0
pip==23.2
pipenv==2022.12.19
PyAudio==0.2.13
pydub==0.25.1
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.15
PyPDF2==3.0.1
pytesseract==0.3.10
python-docx==0.8.11
requests==2.28.2
SpeechRecognition==3.9.0

